Re-unify spindle.h
